Move JsonConvert into apartment.json
