+++ src/main/resources/io/github/oliviercailloux/y2018/apartments/apartment/json/defaultApartments.json
