+++ src/test/resources/io/github/oliviercailloux/y2018/apartments/apartment/json/jsonApartments.json
